[FIX] base_tier_validation: amend authors
I just realized that a new author was added to the module during an
improvement that merged the functionality of base_tier_validation_waiting.
This was overlook during my review and not explicitly accepted. There
have been many contributors adding new functionalities and this one
was not specially big to be included as author.

[BOT] post-merge updates